refactor: extract general settings section
