+++ src/components/Settings.jsx
@@ -1,10 +1,11 @@
 import { useState, useEffect } from 'react'
-import { Form, Input, InputNumber, Button, Switch, message, Divider, Tag, Space, Select, Checkbox, Anchor, Radio, Modal, Tooltip } from 'antd'
+import { Form, Input, InputNumber, Button, Switch, message, Divider, Tag, Space, Select, Checkbox, Anchor, Tooltip } from 'antd'
 import { MailOutlined, LockOutlined, GlobalOutlined, FolderOutlined, QuestionCircleOutlined } from '@ant-design/icons'
 import { mailApi, USE_MOCK } from '../services/api'
 import { getSettings, saveSettings, formatFolderName } from '../services/settings'
 import AiSettingsSection from './settings/AiSettingsSection'
 import ArchiveSettingsSection from './settings/ArchiveSettingsSection'
+import GeneralSettingsSection from './settings/GeneralSettingsSection'
 import SopSettingsSection from './settings/SopSettingsSection'
 import './Settings.css'
 
@@ -126,36 +127,6 @@ function Settings() {
     }
   }
 
-  const handleWindowStyleChange = async (e) => {
-    const newStyle = e.target.value
-    updateSetting('windowStyle', newStyle)
-
-    // 通知Electron主进程存储该设置
-    if (window.electronAPI?.saveSetting) {
-      await window.electronAPI.saveSetting('windowStyle', newStyle)
-    }
-
-    Modal.confirm({
-      title: '重启生效',
-      content: '窗口样式的更改需要重启应用后才会生效，是否立即重启？',
-      okText: '立即重启',
-      cancelText: '稍后重启',
-      onOk: () => {
-        if (window.electronAPI?.restartApp) {
-          window.electronAPI.restartApp()
-        }
-      }
-    })
-  }
-
-  const handlePreviewUpdatesChange = async (checked) => {
-    updateSetting('enablePreviewUpdates', checked)
-
-    if (window.electronAPI?.saveSetting) {
-      await window.electronAPI.saveSetting('enablePreviewUpdates', checked)
-    }
-  }
-
   const handleSelectFolder = async () => {
     if (window.electronAPI?.selectFolder) {
       const selectedPath = await window.electronAPI.selectFolder()
@@ -174,57 +145,7 @@ function Settings() {
   return (
     <div className="settings-container">
       <div className="settings-content">
-        {/* 常规设置 */}
-        <div id="general-settings" className="settings-block">
-          <h2>常规设置</h2>
-          <div className="settings-section">
-            <div className="section-header">
-              <h3>运行模式</h3>
-            </div>
-            <div className="mode-status">
-              {USE_MOCK ? (
-                <Tag color="orange">Mock 模式（外网开发）</Tag>
-              ) : (
-                <Tag color="green">已连接邮件服务器</Tag>
-              )}
-              <p className="mode-hint">
-                {USE_MOCK
-                  ? 'Mock 模式下使用模拟邮件数据，但文件夹创建为真实操作'
-                  : '当前连接真实邮件服务器'}
-              </p>
-            </div>
-          </div>
-          <div className="settings-section" style={{ marginTop: 24 }}>
-            <div className="section-header">
-              <h3>窗口样式</h3>
-            </div>
-            <Radio.Group
-              onChange={handleWindowStyleChange}
-              value={settings.windowStyle}
-              optionType="button"
-              buttonStyle="solid"
-            >
-              <Radio.Button value="integrated" style={{ width: 120, textAlign: 'center' }}>一体化</Radio.Button>
-              <Radio.Button value="classic" style={{ width: 120, textAlign: 'center' }}>经典</Radio.Button>
-            </Radio.Group>
-          </div>
-
-          <div className="settings-section" style={{ marginTop: 24 }}>
-            <div className="section-header">
-              <h3>更新通道</h3>
-            </div>
-            <div className="setting-item inline">
-              <label>启用预览版更新</label>
-              <Switch
-                checked={!!settings.enablePreviewUpdates}
-                onChange={handlePreviewUpdatesChange}
-              />
-            </div>
-            <p className="setting-hint">
-              关闭时只跟踪正式版；开启后会允许检查 alpha/preview 预览版。
-            </p>
-          </div>
-        </div>
+        <GeneralSettingsSection settings={settings} onSettingsChange={setSettings} />
 
         <Divider />
